changed to var type long

+++ learning_java/improved_fibonacci.java
@@ -5,8 +5,8 @@ class improved_fibonacci {
      * Mark the evens with a *
      */
     public static void main(String[] args){
-        int lo = 1;
-        int hi =1;
+        long lo = 1;
+        long hi =1;
         String mark;
         System.out.println("1: " + lo);
         for (int i = 2; i <= MAX_INDEX; i++) {
